refactor: split component
